backend/requirements.txt
fastapi
passlib[bcrypt]
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
python-jose[cryptography]
python-multipart
slowapi
sqlalchemy
#starlette
uvicorn[standard]